+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file. This file i
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.74] - 2026-03-07
+
+### Removed
+
+- Removed `.integrity-suite/docs/prompt.md`, `.integrity-suite/docs/workflow.md`, and `.integrity-suite/docs/requirements.md` to simplify process.
+- Updated testing suite logic to omit validating requirements completely.
+
 ## [1.4.73] - 2026-03-07
 
 ### Changed